requirements.txt
fastapi[standard]
psycopg2-binary
SQLAlchemy
python-dotenv
pyjwt